[pipelines] making fund settlement a separate action
