+++ .buildkite/close-settlements.yml
@@ -2,7 +2,8 @@ agents:
   queue: "snapshots"
 
 steps:
-  - command: echo "--> Start of concurrency gate"
+  - label: ":closed_lock_with_key: Concurrency gate lock"
+    command: echo "--> Start of concurrency gate"
     concurrency_group: 'validator-bonds/close-settlements'
     concurrency: 1
 
@@ -58,18 +59,18 @@ steps:
 
   - wait: ~
 
-  - label: ":campfire: Close settlements"
+  - label: ":campfire::arrow_right: Close settlements"
     env:
       RUST_LOG: info,solana_transaction_builder_executor=debug,solana_transaction_builder=debug,solana_transaction_executor=debug,close_settlement=debug
       # RUST_BACKTRACE: full
     commands:
     - '. "$HOME/.cargo/env"'
     - 'prior_build_number=$(($$BUILDKITE_RETRY_COUNT - 1))'
-    - 'buildkite-agent artifact download --include-retried-jobs build-report.$$prior_build_number . || true'
+    - 'buildkite-agent artifact download --include-retried-jobs close-execution-report.$$prior_build_number . || true'
     - |
-      cp "build-report.$$prior_build_number" "build-report.$$BUILDKITE_RETRY_COUNT" || true
-      rm -f "build-report.$$prior_build_number"
-      echo "#ATTEMPT $$BUILDKITE_RETRY_COUNT" | tee -a "./build-report.$$BUILDKITE_RETRY_COUNT"
+      cp "close-execution-report.$$prior_build_number" "close-execution-report.$$BUILDKITE_RETRY_COUNT" || true
+      rm -f "close-execution-report.$$prior_build_number"
+      echo "#ATTEMPT $$BUILDKITE_RETRY_COUNT" | tee -a "./close-execution-report.$$BUILDKITE_RETRY_COUNT"
     - 'buildkite-agent artifact download --include-retried-jobs target/release/close-settlement .'
     - 'buildkite-agent artifact download --include-retried-jobs past-settlements.json .'
     - 'chmod +x target/release/close-settlement'
@@ -80,21 +81,21 @@ steps:
         --operator-authority "$$VALIDATOR_BONDS_OPERATOR_AUTHORITY" \
         --fee-payer "$$VALIDATOR_BONDS_SETTLEMENTS_WALLET" \
         --marinade-wallet "$$VALIDATOR_BONDS_FUNDING_WALLET" \
-        --past-settlements ./past-settlements.json | tee -a "./build-report.$$BUILDKITE_RETRY_COUNT"
+        --past-settlements ./past-settlements.json | tee -a "./close-execution-report.$$BUILDKITE_RETRY_COUNT"
     key: 'close-settlement'
     artifact_paths:
-      - "./build-report.*"
+      - "./close-execution-report.*"
     retry:
       automatic:
         - exit_status: 100
           limit: 5
 
-  - label: "Notification setup"
+  - label: ":memo: Notification setup: Close Settlements"
     env:
       gs_bucket: gs://marinade-validator-bonds-mainnet
     commands:
      - 'mkdir ./reports'
-     - 'buildkite-agent artifact download --include-retried-jobs "build-report.*" ./reports/ || echo "No report ERROR" > ./reports/build-report.error'
+     - 'buildkite-agent artifact download --include-retried-jobs "close-execution-report.*" ./reports/ || echo "No report ERROR" > ./reports/close-execution-report.error'
      - 'report_path="./reports/$(ls -v1 reports/ | tail -n 1)"'
      - 'cp "$$report_path" ./close-report.txt'
      - 'buildkite-agent meta-data set attempts_count "$(grep -c ATTEMPT ./close-report.txt)"'
@@ -134,6 +135,7 @@ steps:
 
   - wait: ~
 
-  - command: echo "End of concurrency gate <--"
+  - label: ":unlock: Concurrency gate unlock"
+    command: echo "End of concurrency gate <--"
     concurrency_group: 'validator-bonds/close-settlements'
     concurrency: 1
