+++ .buildkite/verify-settlements.yml
@@ -12,7 +12,8 @@ agents:
 #  * programs/validator-bonds/src/instructions/settlement/cancel_settlement.rs
 
 steps:
-  - command: echo "--> Start of concurrency gate"
+  - label: ":closed_lock_with_key: Concurrency gate lock"
+    command: echo "--> Start of concurrency gate"
     concurrency_group: 'validator-bonds/verify-settlements'
     concurrency: 1
 
@@ -60,14 +61,14 @@ steps:
 
   - wait: ~
 
-  - label: ":campfire: Verify settlements"
+  - label: ":campfire::arrow_right: Verify settlements"
     env:
       RUST_LOG: info,solana_transaction_builder_executor=debug,solana_transaction_builder=debug,solana_transaction_executor=debug,verify_settlement=debug
     commands:
     - '. "$HOME/.cargo/env"'
     - |
       
-      echo "#ATTEMPT $((BUILDKITE_RETRY_COUNT+1))" | tee -a "./build-report.$$BUILDKITE_RETRY_COUNT"
+      echo "#ATTEMPT $((BUILDKITE_RETRY_COUNT+1))" | tee -a "./verify-execution-report.$$BUILDKITE_RETRY_COUNT"
     - 'buildkite-agent artifact download --include-retried-jobs target/release/verify-settlement .'
     - 'buildkite-agent artifact download --include-retried-jobs past-settlements.json .'
     - 'chmod +x target/release/verify-settlement'
@@ -75,19 +76,19 @@ steps:
       set -o pipefail
       ./target/release/verify-settlement \
         --rpc-url $$RPC_URL \
-        --past-settlements ./past-settlements.json | tee -a "./build-report.$$BUILDKITE_RETRY_COUNT"
+        --past-settlements ./past-settlements.json | tee -a "./verify-execution-report.$$BUILDKITE_RETRY_COUNT"
     key: 'verify-settlement'
     artifact_paths:
-      - "./build-report.*"
+      - "./verify-execution-report.*"
     retry:
       automatic:
         - exit_status: 100
           limit: 3
 
-  - label: "Notification setup"
+  - label: ":memo: Notification setup: Verify Settlements"
     commands:
      - 'mkdir ./reports'
-     - 'buildkite-agent artifact download --include-retried-jobs "build-report.*" ./reports/ || echo "[\"No report ERROR\"]" > ./reports/build-report.error'
+     - 'buildkite-agent artifact download --include-retried-jobs "verify-execution-report.*" ./reports/ || echo "[\"No report ERROR\"]" > ./reports/verify-execution-report.error'
      - 'report_path="./reports/$(ls -v1 reports/ | tail -n 1)"'
      - |
        ATTEMPT_COUNT=$(grep -c ATTEMPT "$$report_path" | sed 's/.*ATTEMPT //g')
@@ -139,6 +140,7 @@ steps:
 
   - wait: ~
 
-  - command: echo "End of concurrency gate <--"
+  - label: ":unlock: Concurrency gate unlock"
+    command: echo "End of concurrency gate <--"
     concurrency_group: 'validator-bonds/verify-settlements'
     concurrency: 1
